requirements.txt
Flask==2.1.0
requests==2.31.0
google-cloud-datastore==2.3.0
environs==9.5.0
click==8.1.7
pytest==7.4.2